# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AI Agent API
  description: |
    Go AI Agent 后端接口（Gin + DeepSeek + PostgreSQL/pgvector）。

    - 业务接口需 Header：`X-API-Key`
    - 可选 Header：`X-Request-ID`（缺省由服务端生成）
    - 流式接口返回 `text/event-stream`（SSE）
  version: 1.0.0
  contact:
    name: ai-agent
servers:
  - url: http://localhost:18090
    description: 本地开发
...
paths:
...
  /api/v1/corpora/{id}/documents:
    post:
      tags: [Corpus]
      summary: 上传/写入文档并索引
      description: |
        支持 JSON 正文，或 `multipart/form-data` 上传 `file`。

        文件类型：
        - 文本：txt/md/csv/json/html…
        - Word：docx（不支持旧版 doc）
        - PDF：优先提取文字；扫描件走 OCR（需 tesseract + pdftoppm）
        - 图片：png/jpg/jpeg/webp/bmp/tif/gif（Tesseract OCR）

        详见 docs/EXTRACT.md。
      operationId: addDocument
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/RequestIdHeader"
        - $ref: "#/components/parameters/UUIDPathId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddDocumentJSONRequest"
            example:
              title: sla.md
              content: "# SLA\n响应时间 5 分钟"
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: 文档已创建（可能仍为 indexed/failed）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
    get:
      tags: [Corpus]
      summary: 文档列表
      operationId: listDocuments
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/RequestIdHeader"
        - $ref: "#/components/parameters/UUIDPathId"
      responses:
        "200":
          description: 列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Document"
  /api/v1/corpora/{id}/documents/{doc_id}:
    delete:
      tags: [Corpus]
      summary: 删除文档
      operationId: deleteDocument
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/RequestIdHeader"
        - $ref: "#/components/parameters/UUIDPathId"
        - name: doc_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: 已删除
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: 配置中的业务 API Key（示例 change-me-api-key）
  parameters:
    RequestIdHeader:
      name: X-Request-ID
      in: header
      required: false
      schema:
        type: string
        format: uuid
      description: 可选；缺省由服务端生成
    UUIDPathId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: 参数错误
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalError:
      description: 内部错误
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: unauthorized
            message:
              type: string
              example: invalid api key
        request_id:
          type: string
          format: uuid
...
    AddDocumentJSONRequest:
      type: object
      required: [content]
      properties:
        title:
          type: string
        content:
          type: string
    Document:
      type: object
      properties:
        id:
          type: string
          format: uuid
        corpus_id:
          type: string
          format: uuid
        title:
          type: string
        source:
          type: string
        content_hash:
          type: string
        status:
          type: string
          enum: [pending, indexed, failed]
        error_message:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...